requirements.txt
torch>=1.13.0
torchvision>=0.14.0
transformers>=4.21.0
opencv-python>=4.7.0
Pillow>=9.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
numpy>=1.21.0
seaborn>=0.11.0
timm>=0.6.0
PyYAML>=6.0
click>=8.0.0
requests
tqdm
